+++ gDRutils/R/identifiers_list.R
@@ -12,7 +12,7 @@ IDENTIFIERS_LIST <- list(
   cellline_subtype = "subtype",
 
   drug = "Gnumber",
-  drugname = "DrugName",
+  drug_name = "DrugName",
   drug_moa = "drug_moa",
   # corresponds to the field 'gcsi_drug_name' from gCellGenomics::getDrugs()
 
@@ -29,16 +29,52 @@ IDENTIFIERS_LIST <- list(
  
   # ids for the 2nd drug 
   drug2 = "Gnumber_2",
-  drugname2 = "DrugName_2",
+  drug_name2 = "DrugName_2",
   drug_moa2 = "drug_moa_2",
   concentration2 = "Concentration_2",
   
   # ids for the 3rd drug 
   drug3 = "Gnumber_3",
-  drugname3 = "DrugName_3",
+  drug_name3 = "DrugName_3",
   drug_moa3 = "drug_moa_3",
   concentration3 = "Concentration_3",
   
   # data source
   data_source = "data_source"
+)
+
+REQ_COL_IDENTIFIERS <- c(
+  "duration",
+  "cellline",
+  "cellline_name",
+  "drug",
+  "drug_name",
+  "masked_tag",
+  "concentration"
+)
+
+EXPECT_ONE_IDENTIFIERS <- c(
+  "duration",
+  "cellline",
+  "cellline_name",
+  "cellline_tissue",
+  "cellline_ref_div_time",
+  "cellline_parental_identifier",
+  "cellline_subtype",
+  "drug",
+  "drug_name",
+  "drug_moa",
+  "masked_tag",
+  "concentration",
+  "template",
+  "barcode",
+  "drug2",
+  "drug_name2",
+  "drug_moa2",
+  "concentration2",
+  "drug3",
+  "drug_name3",
+  "drug_moa3",
+  "concentration3",
+  "data_source"
 )