+++ gDRutils/R/utils.R
@@ -52,6 +52,7 @@ extend_normalization_type_name <- function(x) {
 #' 
 #' @param x charvec expected subset
 #' @param choices charvec reference set
+#' @param ... Additional arguments to pass to \code{checkmate::test_choice}
 #' @export
 assert_choices <- function(x, choices, ...) {
   out <- vapply(x, function(y) {
@@ -66,5 +67,27 @@ assert_choices <- function(x, choices, ...) {
         toString(choices)
       )
     stop(msg)
+  }
+}
+
+#' Lapply through all the experiments in MultiAssayExperiment object
+#' 
+#' @param mae MultiAssayExperiment object
+#' @param FUN function that should be applied on each experiment of MultiAssayExperiment object
+#' @param unify logical indicating if the output should be a unlisted object of unique
+#' values across all the experiments 
+#' @export
+#' 
+#' @author Bartosz Czech <bartosz.czech@@contractors.roche.com>
+MAEpply <- function(mae, FUN, unify = FALSE) {
+  checkmate::assert_class(mae, "MultiAssayExperiment")
+  checkmate::assert_function(FUN)
+  checkmate::assert_flag(unify)
+  experiments <- as.list(MultiAssayExperiment::experiments(mae))
+  out <- lapply(experiments, FUN)
+  if (unify) {
+    unique(unlist(out))
+  } else {
+    out
   }
 }